Soften the language statement
This is just a suggestion.  I think it would be good not to be in a 'mostly Python' corner, and to make it plain that we'd take, say, Julia.

BTW there are now multiple mentions of Matlab.  Isn't that a conflict with [item 3 on the FAQ](https://github.com/ReScience/rescience.github.io/blob/master/05-FAQ.md#can-i-submit-a-paper-using-proprietary-tools-such-as-matlab-) ?